+++ spec/appraisal/gem_manager/base_spec.rb
@@ -3,9 +3,10 @@
 require "appraisal/gem_manager/base"
 
 RSpec.describe Appraisal::GemManager::Base do
+  subject(:base) { described_class.new(gemfile_path, project_root) }
+
   let(:gemfile_path) { "/path/to/gemfile" }
   let(:project_root) { Pathname.new("/path/to/project") }
-  subject(:base) { described_class.new(gemfile_path, project_root) }
 
   describe "#initialize" do
     it "stores gemfile_path" do
@@ -21,7 +22,7 @@ RSpec.describe Appraisal::GemManager::Base do
     it "raises NotImplementedError" do
       expect { base.install }.to raise_error(
         NotImplementedError,
-        "Appraisal::GemManager::Base#install must be implemented"
+        "Appraisal::GemManager::Base#install must be implemented",
       )
     end
   end
@@ -30,7 +31,7 @@ RSpec.describe Appraisal::GemManager::Base do
     it "raises NotImplementedError" do
       expect { base.update }.to raise_error(
         NotImplementedError,
-        "Appraisal::GemManager::Base#update must be implemented"
+        "Appraisal::GemManager::Base#update must be implemented",
       )
     end
   end
@@ -39,7 +40,7 @@ RSpec.describe Appraisal::GemManager::Base do
     it "raises NotImplementedError" do
       expect { base.name }.to raise_error(
         NotImplementedError,
-        "Appraisal::GemManager::Base#name must be implemented"
+        "Appraisal::GemManager::Base#name must be implemented",
       )
     end
   end
@@ -48,7 +49,7 @@ RSpec.describe Appraisal::GemManager::Base do
     it "raises NotImplementedError" do
       expect { base.available? }.to raise_error(
         NotImplementedError,
-        "Appraisal::GemManager::Base#available? must be implemented"
+        "Appraisal::GemManager::Base#available? must be implemented",
       )
     end
   end